Keep take audio bounded across compaction
Keep speech ranges stable across long recordings and source compaction. Count lifetime input independently from retained audio, enforce one sample budget across resident, queued, and decoding owners, and transfer ownership before draining the rolling buffer.

- `RetainedPcmBudget` shares one exact reservation across `RollingPcm`, `RetainedPcm`, and `RetainedPcmOwner`. Transfers move ownership before compaction, while copies reserve additional capacity.
- `with_lifetime_guard` moves the retained-audio owner into each worker `Job`. Epoch cancellation cannot release that reservation before the worker retires the request.
- `AudioBuffer` keeps lifetime input as `u64` across repeated output drains. `SegmentOutcome`, `FinalRangeOutcome`, and `AcceptedHypothesis` keep absolute ranges after the resident buffer origin advances.
- `MAX_PENDING_FINAL_OUTCOMES` and `MAX_PENDING_ACCEPTED_HYPOTHESES` impose visible fixed limits on retained text history. `commit` restores the input and delays mailbox reservation when sealing exceeds the shared audio budget.
- `blocked_interim_keeps_exact_budget_until_worker_retirement_and_commit_retries` and `blocked_final_keeps_budget_after_epoch_cancellation_until_worker_retirement` pin reservation ownership through cancellation, retry, and worker retirement.
- `crates/gateway-stt/src/segment.rs` adds no forced final window or overlap reconciliation.

Design: extends parameter-object @ crates/gateway-stt-engine/src/decoder.rs::DecodeRequest boundary: pub
Design: new surface-growth @ crates/gateway-stt-engine/src/decoder.rs::DecodeRequest::with_lifetime_guard boundary: pub
Design: new newtype @ crates/gateway-stt-engine/src/decoder.rs::RequestLifetime
Design: extends shared-mutable-state @ crates/gateway-stt-engine/src/worker.rs::Transcriber
Design: extends shotgun-surgery @ crates/gateway-stt/src
Design: extends oversized-unit @ crates/gateway-stt/src/audio.rs
Design: extends value-object @ crates/gateway-stt/src/audio.rs::CommittedAudio
Design: extends oversized-unit @ crates/gateway-stt/src/realtime/input.rs
Design: new shared-parameter-cluster @ crates/gateway-stt/src/realtime/input.rs::UncommittedInput::first_append_with_pcm_limit
Design: new shared-parameter-cluster @ crates/gateway-stt/src/realtime/input.rs::UncommittedInput::from_audio_and_take
Design: extends oversized-unit @ crates/gateway-stt/src/realtime/result_mailbox.rs
Design: extends oversized-unit @ crates/gateway-stt/src/realtime/session/items.rs
Design: new oversized-unit @ crates/gateway-stt/src/realtime/session/items/tests.rs
Design: extends oversized-unit @ crates/gateway-stt/src/realtime/session/route.rs
Design: extends pure-function @ crates/gateway-stt/src/realtime/session/route.rs::sample_millis deps: u64
Design: extends oversized-unit @ crates/gateway-stt/src/realtime/session/state.rs
Design: extends value-object @ crates/gateway-stt/src/segment.rs::SegmentOutcome
Design: extends oversized-unit @ crates/gateway-stt/src/take.rs
Design: extends parameter-object @ crates/gateway-stt/src/take.rs::Take
Design: extends shared-parameter-cluster @ crates/gateway-stt/src/take.rs::Take::next_window_snapshot
Design: extends oversized-unit @ crates/gateway-stt/src/take/final_outcome.rs
Design: extends value-object @ crates/gateway-stt/src/take/final_outcome.rs::FinalRangeOutcome
Design: extends pure-function @ crates/gateway-stt/src/take/final_outcome.rs::assemble_completion deps: &[AcceptedHypothesis],&[FinalRangeOutcome],u64
Design: extends pure-function @ crates/gateway-stt/src/take/final_outcome.rs::skipped_outcomes_exactly_cover deps: &Range<u64>,&[FinalRangeOutcome],usize
Design: extends oversized-unit @ crates/gateway-stt/src/take/finalization.rs
Design: extends message-passing @ crates/gateway-stt/src/take/finalization.rs::FinalPipeline
Design: extends shared-mutable-state @ crates/gateway-stt/src/take/finalization.rs::FinalPipeline
Design: new newtype @ crates/gateway-stt/src/take/finalization.rs::FinalSegmentOwner
Design: new oversized-unit @ crates/gateway-stt/src/take/pcm.rs
Design: new shared-mutable-state @ crates/gateway-stt/src/take/pcm.rs::RetainedPcmBudget
Design: new newtype @ crates/gateway-stt/src/take/pcm.rs::RetainedPcmBudget
Design: extends oversized-unit @ crates/gateway-stt/src/take/state.rs
Design: extends shared-mutable-state @ crates/gateway-stt/src/take/state.rs::TakeState
Design: extends oversized-unit @ crates/gateway-stt/src/take/window.rs
Design: extends value-object @ crates/gateway-stt/src/take/window.rs::AcceptedHypothesis
Design: extends shared-parameter-cluster @ crates/gateway-stt/src/take/window.rs::WholeWindowState::next
Design: new shared-parameter-cluster @ crates/gateway-stt/src/take/window.rs::WholeWindowState::try_next
Design: extends pure-function @ crates/gateway-stt/src/test_fixtures/segment.rs::segment_ranges deps: &[f32] boundary: pub
Design: extends surface-growth @ crates/gateway-stt/src/test_fixtures/segment.rs::segment_ranges boundary: pub
Violates: A2 - credential ownership in crates/gateway-stt/src/take is not determinable from diff
Pending: N6 - compounds
Pending: N60 - compounds
Deferred: forced final windows and overlap reconciliation remain absent
Plan: vibe/2026-09-07-1-promptforge-debt.md

+++ crates/gateway-stt/src/realtime/session/state.rs
@@ -19,9 +19,9 @@ pub(super) enum InterimTaskOutput {
     Decode {
         epoch: InterimEpoch,
         item_id: String,
-        segment_start: usize,
-        audio_start: usize,
-        audio_end: usize,
+        segment_start: u64,
+        audio_start: u64,
+        audio_end: u64,
         transcript: Result<String, String>,
     },
 }
@@ -72,7 +72,7 @@ pub(crate) struct Session {
     pub(super) current_epoch: Option<InterimEpoch>,
     pub(super) next_epoch: u64,
     pub(super) interim_task: Option<InterimTask>,
-    pub(super) last_interim_window: Option<(usize, usize, usize)>,
+    pub(super) last_interim_window: Option<(u64, u64, u64)>,
     pub(super) canceled_tasks: Vec<InterimTask>,
     pub(super) canceled_task_failed: bool,
     pub(super) committed: HashMap<String, CommittedItem>,
